requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
aiohttp=3.7.4.post0=py37h5e8e339_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py37h89c1867_0
chardet=4.0.0=py37h89c1867_1
charset-normalizer=2.0.3=pypi_0
click=8.0.1=py37h89c1867_0
dnspython=1.16.0=pypi_0
et-xmlfile=1.1.0=pypi_0
idna=3.2=pypi_0
importlib-metadata=4.6.1=py37h89c1867_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
motor=2.3.0=pyhd8ed1ab_0
multidict=5.1.0=py37h5e8e339_1
ncurses=6.2=he6710b0_1
numpy=1.21.0=pypi_0
openai=0.10.1=pypi_0
openpyxl=3.0.7=pypi_0
openssl=1.1.1k=h7f98852_0
pandas=1.3.0=pypi_0
pandas-stubs=1.2.0.1=pypi_0
pip=21.1.3=py37h06a4308_0
pymongo=3.12.0=py37hcd2ae1e_0
python=3.7.10=h12debd9_4
python-dateutil=2.8.2=pypi_0
python-dotenv=0.18.0=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2021.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
setuptools=52.0.0=py37h06a4308_0
six=1.16.0=pypi_0
slack-bolt=1.7.0=pypi_0
slack-sdk=3.8.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.10=hbc83047_0
tornado=6.1=py37h5e8e339_1
tqdm=4.61.2=pypi_0
typing-extensions=3.10.0.0=hd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
urllib3=1.26.6=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py37h5e8e339_2
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
